Classify BinaryHeap & LinkedList unit tests as such
