+++ library/alloc/src/collections/binary_heap.rs
@@ -155,6 +155,9 @@ use crate::vec::{self, AsIntoIter, Vec};
 
 use super::SpecExtend;
 
+#[cfg(test)]
+mod tests;
+
 /// A priority queue implemented with a binary heap.
 ///
 /// This will be a max-heap.
