+++ library/alloc/src/collections/binary_heap/tests.rs
@@ -1,5 +1,5 @@
-use std::collections::binary_heap::{Drain, PeekMut};
-use std::collections::BinaryHeap;
+use super::*;
+use crate::boxed::Box;
 use std::iter::TrustedLen;
 use std::panic::{catch_unwind, AssertUnwindSafe};
 use std::sync::atomic::{AtomicU32, Ordering};
